requirements3.txt
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
checkbox-support==0.22
click==8.0.3
cmake==3.26.4
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
emoji==1.6.1
entrypoints==0.4
et-xmlfile==1.1.0
fasteners==0.14.1
feedparser==6.0.8
filelock==3.12.1
Flask==2.0.2
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
guacamole==0.9.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.0.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.5.post0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.1
nltk==3.6.5
notebook==6.4.8
numpy==1.21.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
openpyxl==3.0.9
packaging==21.3
padme==1.1.1
pandas==1.3.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plainbox==0.25
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.19.1
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
PyAudio==0.2.11
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pythran==0.10.0
pytz==2021.3
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.11.2
reportlab==3.6.8
requests==2.26.0
scikit-learn==1.0
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.1b0
sentiment-analysis-spanish==0.0.25
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
stanza==1.2.3
statsmodels==0.13.0
sympy==1.9
system-service==0.3
systemd-python==234
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.0.0
torch==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
triton==2.0.0
typing-extensions==3.10.0.2
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unity-lens-photos==1.0
unity-scope-calculator==0.1
unity-scope-devhelp==0.1
unity-scope-manpages==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-zotero==0.1
urllib3==1.26.7
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==2.0.0
wordcloud==1.9.2
xdg==5
xdiagnose==3.8.10
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.2
zipp==1.0.0
